♻️ Refactor: 헤더 컴포넌트로 추출
